requirements.txt
about-time==3.1.1
alive-progress==2.4.1
grapheme==0.6.0
numpy==1.23.4
pandas==1.5.1
progress==1.6
progressbar2==4.2.0
python-dateutil==2.8.2
python-utils==3.4.5
pytz==2022.6
six==1.16.0
tqdm==4.64.1
